moved app code to a separate app file
